Miscellaneous updates
rustc is now happy to let you elide the static lifetime when declaring a
static variable, so that's nice. rustfmt has been updated to be less
egregious (read: they have adopted my preferred format) for long arrays,
so that is also nice.

And I have changed my mind about using typedefs for result values, so
I've changed that.

Also, the error struct was named inappropriately. An out of range error
on creating a new Roman value is not a parsing error. To reflect this,
ParseRomanError is now just called RomanError.