Fix y4k bug
Apparently, 4999 is the largest accepted, commonly-expressible Roman numeral. The library previously had this value as 3999.

I attempted to just raise the hard-coded limit to 4999, but I discovered that even though the printing function was able to spit out the correct MCMCMCFUCK for 4999, the parsing function was apt to choke on it.

The new parsing function has a little more abstraction thrown in and behaves more predictably. The overall strategy (based on numeral units, if you recall) is unchanged, and it still reads everything in one shot, left to right.