Don't use ASCII trait
Rust 1.23 stabilizes the move of the ASCII trait methods from the trait
to the pertinent types themselves. Of course, this is
backward-incompatible (?!) I guess, with versions of the rust compiler
that don't have that feature in their version of the standard library
(?!) (Hell if I know, shut up.), so in order to continue to support
them, I would need to leave the trait in there but mark it with an
allow(unused) config flag...

Screw that.

We now or the byte with 32 in order to ensure that we're testing a
"lowercase" value.